refactor: hexagonal and clean architecture
