+++ src/core/configs/otp.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class OTPConfig:
+    expires_in: int = 300
