+++ src/core/ports/otp.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class OTPGeneratorPort(ABC):
+    @abstractmethod
+    def generate_code(self, moving_factor: int) -> str:
+        pass
