+++ src/infra/adapters/otp.py
@@ -1,13 +1,13 @@
 from pyotp import OTP as PyOTP
-from src.app.configs.oath import OATH_CONFIG
+from src.core.ports.otp import OTPGeneratorPort
+from src.infra.adapters.shared import default_hash_method
 
 
-class OTPGenerator:
+class OTPGenerator(OTPGeneratorPort):
     def __init__(self, raw_secret: str):
-        self.raw_secret = raw_secret
         self.server = PyOTP(
-            self.raw_secret,
-            digest=OATH_CONFIG.hash_method
+            raw_secret,
+            digest=default_hash_method
         )
 
     def generate_code(self, moving_factor: int) -> str:
