+++ src/infra/adapters/shared.py
@@ -1,0 +1,4 @@
+from hashlib import sha256
+from typing import Callable
+
+default_hash_method: Callable = sha256
