+++ src/tests/integration/services/test_hotp.py
@@ -2,7 +2,7 @@ from re import search
 from pyotp import HOTP
 from flask.testing import FlaskClient
 
-from src.app.configs.oath import OATH_CONFIG
+from src.infra.adapters.shared import default_hash_method
 
 
 def test_hotp_create_and_verify_request_200_twice_success(
@@ -14,7 +14,7 @@ def test_hotp_create_and_verify_request_200_twice_success(
     hotp_uri = create_response.json["uri"]
     secret_match = search(r"secret=([A-Z2-7]+)", hotp_uri)
     secret = secret_match.group(1)
-    hotp = HOTP(secret, digest=OATH_CONFIG.hash_method)
+    hotp = HOTP(secret, digest=default_hash_method)
     current_otp = hotp.at(0)
 
     verify_response = client.get(f"/api/hotp?otp={current_otp}")
@@ -37,7 +37,7 @@ def test_hotp_create_and_verify_request_200_twice_success_with_resync_protocol(
     hotp_uri = create_response.json["uri"]
     secret_match = search(r"secret=([A-Z2-7]+)", hotp_uri)
     secret = secret_match.group(1)
-    client_hotp = HOTP(secret, digest=OATH_CONFIG.hash_method)
+    client_hotp = HOTP(secret, digest=default_hash_method)
     desynchronized_otp = client_hotp.at(35)
 
     verify_response = client.get(
@@ -74,7 +74,7 @@ def test_hotp_create_and_verify_wrong_otp_200_failure_with_resync_protocol(
     hotp_uri = create_response.json["uri"]
     secret_match = search(r"secret=([A-Z2-7]+)", hotp_uri)
     secret = secret_match.group(1)
-    client_hotp = HOTP(secret, digest=OATH_CONFIG.hash_method)
+    client_hotp = HOTP(secret, digest=default_hash_method)
     otp_out_of_the_resync_range = client_hotp.at(6)
 
     verify_response = client.get(
@@ -95,7 +95,7 @@ def test_hotp_create_and_verify_already_used_otp_200_failure(
     hotp_uri = create_response.json["uri"]
     secret_match = search(r"secret=([A-Z2-7]+)", hotp_uri)
     secret = secret_match.group(1)
-    hotp = HOTP(secret, digest=OATH_CONFIG.hash_method)
+    hotp = HOTP(secret, digest=default_hash_method)
     current_otp = hotp.at(30)
 
     verify_response = client.get(f"/api/hotp?otp={current_otp}")
