+++ src/tests/unit/core/configs/test_otp.py
@@ -1,0 +1,5 @@
+from src.core.configs.otp import OTPConfig
+
+
+def test_otp_config():
+    assert OTPConfig.expires_in == 300
